+++ tests/testit/test-table.R
@@ -10,4 +10,8 @@ assert(
   'kable() does not discard row names when there is only one row',
   identical(kable(m), c('|id  |  x|  y|', '|:---|--:|--:|', '|a   |  1|  2|'))
 )
+
+assert(
+  'kable() recycles the align argument correctly',
+  identical(kable(m, align = 'c'), c('|id  | x | y |', '|:---|:-:|:-:|', '|a   | 1 | 2 |'))
 )